feat: agregar componente DataTable y funcionalidad de registro de etiquetas
