+++ frontend/src/components/admin/dialogDemo.jsx
@@ -13,41 +13,41 @@ import {
 } from "@/components/ui/dialog";
 import { useEffect, useState } from "react";
 
-export function DialogDemo({ title, fields }) {
-  const [open, setOpen] = useState(false);
-  const methods = useForm();
-  const { reset } = methods;
+export function DialogDemo({ title, fields, onSubmit }) {
+  const [open, setOpen] = useState(false)
+  const methods = useForm({
+    defaultValues: {
+    estado: "No confirmado",
+  },
+  });
+  const { reset } = methods
 
-  // Agrupar pallets por lado
   const agruparPallets = (pallets) => {
-    const izquierda = [],
-      derecha = [];
-
+    const izquierda = [], derecha = []
     Object.entries(pallets || {}).forEach(([key, value]) => {
-      key.startsWith("izq") && izquierda.push(value);
-      key.startsWith("der") && derecha.push(value);
-    });
+      key.startsWith("izq") && izquierda.push(value)
+      key.startsWith("der") && derecha.push(value)
+    })
+    return { izquierda, derecha }
+  }
 
-    return { izquierda, derecha };
-  };
-
-  // ✅ Enviar datos agrupados
-  const onSubmit = (data) => {
-    const { pallets, ...resto } = data;
-    const palletsAgrupados = agruparPallets(pallets);
+  const handleFormSubmit = (data) => {
+    const { pallets, ...resto } = data
+    const palletsAgrupados = agruparPallets(pallets)
 
     const datosFinales = {
       ...resto,
       pallets: palletsAgrupados,
-    };
+    }
 
-    console.log(datosFinales);
-    alert("Formulario enviado");
-  };
+    
+    onSubmit?.(datosFinales)
+    setOpen(false) 
+  }
 
   useEffect(() => {
-    if (open) reset();
-  }, [open, reset]);
+    if (open) reset()
+  }, [open, reset])
 
   return (
     <Dialog open={open} onOpenChange={setOpen}>
@@ -55,18 +55,15 @@ export function DialogDemo({ title, fields }) {
         <Button variant="outline">Crear Formulario</Button>
       </DialogTrigger>
 
-      {/* Modal con layout limpio, ancho y padding */}
       <DialogContent className="w-full !max-w-[95vw] max-h-[90vh] overflow-y-auto bg-white p-0 rounded-2xl shadow-xl">
         <div className="w-full px-4 py-6">
           <FormProvider {...methods}>
-            <form onSubmit={methods.handleSubmit(onSubmit)}>
+            <form onSubmit={methods.handleSubmit(handleFormSubmit)}>
               <DialogHeader>
-                <DialogTitle className="text-3xl text-center mb-8">
-                  {title}
-                </DialogTitle>
+                <DialogTitle className="text-3xl text-center mb-8">{title}</DialogTitle>
               </DialogHeader>
 
-                <RegInputs fields={fields} />
+              <RegInputs fields={fields} />
 
               <DialogFooter className="mt-8 flex justify-end gap-4">
                 <DialogClose asChild>
@@ -79,5 +76,5 @@ export function DialogDemo({ title, fields }) {
         </div>
       </DialogContent>
     </Dialog>
-  );
+  )
 }
