+++ frontend/src/components/admin/etiqueta/fieldsEtiqueta.jsx
@@ -1,26 +1,12 @@
 export const fields = [
-    { name: "exportador", label: "Exportador", type: "text", required: true },
-    { name: "lote", label: "Lote", type: "text", required: true },
-    { name: "categoría", label: "Categoría", type: "text", required: true },
-    { name: "calibre", label: "Calibre", type: "text", required: true },
-    {
-      name: "peso por unidad",
-      label: "Peso por unidad",
-      type: "text",
-      required: true,
-    },
-    {
-      name: "codigo de empaque",
-      label: "Codigo de empaque",
-      type: "text",
-      required: true,
-    },
-    { name: "clp", label: "CLP", type: "text", required: true },
-    { name: "fecha", label: "Fecha", type: "text", required: true },
-    {
-      name: "trazabilidad",
-      label: "Trazabilidad",
-      type: "text",
-      required: true,
-    },
-  ];
+  { name: "exportador", label: "Exportador", type: "text", required: true },
+  { name: "codLote", label: "Lote", type: "text", required: true },
+  { name: "categoria", label: "Categoría", type: "text", required: true },
+  { name: "calibre", label: "Calibre", type: "text", required: true },
+  { name: "pesoUni", label: "Peso por unidad", type: "text", required: true },
+  { name: "codEmp", label: "Código de empaque", type: "text", required: true },
+  { name: "clp", label: "CLP", type: "text", required: true },
+  { name: "fecha", label: "Fecha", type: "text", required: true },
+  { name: "trazabilidad", label: "Trazabilidad", type: "text", required: true },
+]
+
